+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/mcpdirectstudioapp/Platform.kt
@@ -39,7 +39,8 @@ interface Platform {
     fun pasteFromClipboard():String?
     fun copyToClipboard(text:String)
     fun getenv(key:String):String?
-    fun convertYamlToJson(yaml:String):String
+    fun parseOpenAPIDoc(yaml:String, onResponse:(resp: AIPortServiceResponse<OpenAPIServerDoc>)->Unit)
+
     fun httpRequest(usl:String, parameters:Map<String, JsonElement>, onResponse:(resp:String)->Unit)
     fun hstpRequest(usl:String, parameters:Map<String, JsonElement>, onResponse:(resp:String)->Unit)
     fun hstpRequest(usl:String, parameters:String, onResponse:(resp:String)->Unit)